refactor: Socket social event error handling
- remove instances of `errorMessage` variable from `handleSocialEvents` in `handleSocial.js`.
- implement throwing errors on invalid data in event handler `try` scope in `handleSocial.js`.
- move logic for logging error message to the console and `error` event emission to event handler `catch` scope in `handleSocial.js`.
- add `senderId` and `receiverId` to data validation  in `process friend request` event handler in `handleSocial.js`.
- move `friend request processed` emission logic after successful data validation for both accepted and declined responses in `handleSocial.js`.

- Note: This commit aims to improve error handling in `handleSocial.js`  by removing redundant variables, enhancing data validation, and restructuring error logging and event emission for improved clarity and reliability.

+++ backend/controller/socket/handleSocial.js
@@ -2,17 +2,11 @@ import Friend from "../../models/users/friendModel.js";
 
 const handleSocialEvents = (socket, io, userSocketMap) => {
   socket.on("send friend request", async (data) => {
-    let errorMessage = null;
-
     try {
       const { senderId, receiverId, currentStatus } = data;
 
       if (!senderId || !receiverId) {
-        errorMessage = "Please provide valid friend request data";
-
-        console.error(errorMessage);
-        socket.emit("error", { message: errorMessage });
-        return;
+        throw new Error("Please provide valid friend request data");
       }
 
       const existingRequest = await Friend.findOne({
@@ -23,11 +17,7 @@ const handleSocialEvents = (socket, io, userSocketMap) => {
       });
 
       if (existingRequest) {
-        errorMessage = "Friend request already pending";
-
-        console.error(errorMessage);
-        socket.emit("error", { message: errorMessage });
-        return;
+        throw new Error("Friend request already pending");
       }
 
       const newRequest = new Friend({
@@ -41,11 +31,11 @@ const handleSocialEvents = (socket, io, userSocketMap) => {
       socket.emit("friend request sent", newRequest);
 
       const targetSocketId = userSocketMap.get(receiverId);
+
       if (targetSocketId) {
         io.to(targetSocketId).emit("new friend request", newRequest);
-        console.log(`Friend request sent to userId ${targetSocketId}`);
       } else {
-        console.error(`No active socket for userId: ${targetSocketId}`);
+        throw new Error(`No active socket for userId: ${targetSocketId}`);
       }
     } catch (error) {
       console.error(error.message);
@@ -54,24 +44,17 @@ const handleSocialEvents = (socket, io, userSocketMap) => {
   });
 
   socket.on("process friend request", async (data) => {
-    let errorMessage = null;
     try {
       const { senderId, receiverId, userResponse } = data;
 
-      if (!userResponse) {
-        errorMessage = "Please provide valid room data";
-        console.error(errorMessage);
-        socket.emit("error", { message: errorMessage });
-        return;
+      if (!senderId || !receiverId || !userResponse) {
+        throw new Error("Please provide valid room data");
       }
 
       const validStatuses = ["pending", "accepted", "declined"];
 
       if (!validStatuses.includes(userResponse)) {
-        errorMessage = "Invalid friend request status";
-        console.error(errorMessage);
-        socket.emit("error", { errorMessage });
-        return;
+        throw new Error("Invalid friend request status");
       }
 
         const deletedRequest = await Friend.deleteOne({
@@ -94,17 +77,9 @@ const handleSocialEvents = (socket, io, userSocketMap) => {
         });
 
         const targetSocketId = userSocketMap.get(senderId);
-        if (targetSocketId) {
-          io.to(targetSocketId).emit("friend request processed", {
-            senderId,
-            receiverId,
-            status: "declined",
-          });
 
-          console.log(`Friend request declined for userId: '${targetSocketId}`);
-          return;
-        } else {
-          console.error(`No active socket for userId: ${targetSocketId}`);
+        if (!targetSocketId) {
+          throw new Error(`No active socket for userId: ${targetSocketId}`);
         }
       }
 
@@ -115,23 +90,20 @@ const handleSocialEvents = (socket, io, userSocketMap) => {
       );
 
       if (!friendRequest) {
-        errorMessage = "Friend request not found";
-        console.error(errorMessage);
-        socket.emit("error", { errorMessage });
-        return;
+        throw new Error("Friend request not found");
       }
 
       socket.emit("friend request processed", friendRequest);
       const targetSocketId = userSocketMap.get(senderId);
-      if (targetSocketId) {
-        io.to(targetSocketId).emit("friend request processed", friendRequest);
-        console.log(`Friend request sent to userId ${targetSocketId}`);
-      } else {
-        console.log(`No active socket for userId: ${targetSocketId}`);
+
+      if (!targetSocketId) {
+        throw new Error(`No active socket for user: ${targetSocketId}`);
       }
     } catch (error) {
       console.error("Error processing friend request: ", error.message);
-      socket.emit("error", { error });
+      socket.emit("error", { message: error.message });
+    }
+  });
     }
   });
 };
